+++ src/tools/manifestConverter.js
@@ -6,56 +6,19 @@ import os from "os";
 
 const osCurrent = os.platform();
 
-async function attemptToConvert(original) {
-    let newTemplate = template;
-    // Fill Template As Much as Possible
-    for (let idx in newTemplate) {
-        newTemplate[idx] = original[idx];
+async function addOSSpecArguments(original) {
+    console.log('run')
+    if (osCurrent === "win32") {
+        original.arguments.jvm.push(
+            "-XX:HeapDumpPath=MojangTricksIntelDriversForPerformance_javaw.exe_minecraft.exe.heapdump",
+        );
+        original.arguments.jvm.push("-Dos.version=10.0");
+    } else if (osCurrent === "darwin") {
+        // Unsure if needed on newer Mac versions. require hardware to test
+        // Disabling for now
+        //arguments.push("-XstartOnFirstThread")
     }
-    let javaVersion;
-    if (!newTemplate["javaVersion"]) {
-        javaVersion = {component: "jre-legacy"};
-        newTemplate["javaVersion"] = javaVersion;
-    }
-
-    if (
-        !newTemplate["arguments"] ||
-        !newTemplate["arguments"].jvm ||
-        newTemplate["arguments"].jvm.length === 0
-    ) {
-        let template_arguments = [
-            "-Djava.library.path=${natives_directory}",
-            "-Dminecraft.launcher.brand=${launcher_name}",
-            "-Dminecraft.client.jar=${client_jar}",
-            "-Dminecraft.launcher.version=${launcher_version}",
-            "-cp",
-            "${classpath}",
-            "-Xmx${ram}G",
-            "-XX:+UnlockExperimentalVMOptions",
-            "-XX:+UseG1GC",
-            "-XX:G1NewSizePercent=20",
-            "-XX:G1ReservePercent=20",
-            "-XX:MaxGCPauseMillis=50",
-            "-XX:G1HeapRegionSize=32M",
-        ];
-        if (osCurrent === "win32") {
-            template_arguments.push(
-                "-XX:HeapDumpPath=MojangTricksIntelDriversForPerformance_javaw.exe_minecraft.exe.heapdump",
-            );
-            template_arguments.push("-Dos.version=10.0");
-        } else if (osCurrent === "darwin") {
-            //arguments.push("-XstartOnFirstThread")
-        }
-        if (!newTemplate["arguments"]) {
-            newTemplate["arguments"] = {};
-        }
-        newTemplate["arguments"]["jvm"] = template_arguments;
-        if (original.minecraftArguments) {
-            //gameArguments['game'] =
-            newTemplate["arguments"]["game"] = original.minecraftArguments.split(" ");
-        }
-    }
-    return newTemplate;
+    return original;
 }
 
 async function convertAssets(original) {
@@ -128,7 +91,7 @@ async function convertPre16Assets(original) {
 }
 
 export {
-    attemptToConvert,
+    addOSSpecArguments,
     convertAssets,
     convertLegacyAssets,
     convertPre16Assets,